requirements.txt
numpy>=1.24.0
opencv-python>=4.9.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scikit-image>=0.21
loguru>=0.7.0
python-dotenv>=1.0.0
aiogram>=3.4.0
pytest>=8.0.0
